Update patch set 8
Patch Set 8:

> Patch Set 7:
> 
> Sorry, looks like there is also a merge conflict with CL 200867 that will need to be resolved. 😞

This was resolved in Patch Set 7.
Seems now we have another conflict again... not sure how bad it is, but maybe I'll take a look and fix it if we can merge it fast afterwards since this is kind of a pain.

Patch-set: 8
Reviewer: Gerrit User 32909 <32909@62eb7196-b449-3ce5-99f1-c037f21e1705>