Update patch set 14
Patch Set 14:

> Looking at src/cmd/link/internal/ld/config.go in mustLinkExternal, I see that we're still unconditionally forcing external linking for cgo. Would it makes sense to change this to something conditional on GOOS, e.g. use internal linking for ELF systems?

In the general case we must use external linking for any use of cgo outside the standard library.  There are a number of things that the external linker will handle that the Go linker won't, such as global constructors as used in C++ or in C when using the __constructor__ attribute.  (Uses of cgo inside the standard library are OK because we are careful not to use any features that the Go linker does not support.)

Patch-set: 14
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>