cmd/compile: fix liveness for open-coded defer args for infinite loops
Once defined, a stack slot holding an open-coded defer arg should always be marked
live, since it may be used at any time if there is a panic. These stack slots are
typically kept live naturally by the open-defer code inlined at each return/exit point.
However, we need to do extra work to make sure that they are kept live if a
function has an infinite loop (or a panic exit).

The general idea is that we mark each such slot as live in any block that is
dominated by the block in which the slot (autotmp) is defined.

To make this more efficient, I should:
  - do an analysis to find the blocks that cannot reach the exit (including
    BlockExit blocks) and only specially mark the slots for those blocks.
  - or, at the very least, use the sdom() SparseTree, so as to make the dominator
    check more efficient. For that, I will need to export some functions out of
    ssa.

Updates #35277

Change-Id: I7b53c9bd38ba384a3794386dd0eb94e4cbde4eb1