crypto/tls: refactor certificate and signature algorithm logic
This refactors a lot of the certificate support logic to make it cleaner
and reusable where possible. In particular, everything is now routed
through selectSignatureScheme/isSupportedSignatureAlgorithm +
typeAndHashFromSignatureScheme. These changes will make the following
CLs much simpler.

There should be no change in behavior so far (except we now correctly
require signature_algorithms in TLS 1.3 when using a certificate, and we
won't use Ed25519 in TLS 1.2 if the client didn't send
signature_algorithms).

DO NOT SUBMIT: missing tests, in particular TestSignatureSelection needs
to be rewritten in terms of the new functions.

Updates #32426

Change-Id: Ice9dcaea78d204718f661f8d60efdb408ba41577