Update patch set 1
Patch Set 1:

> > Build is still in progress...
> > This change failed on illumos-amd64:
> > See https://storage.googleapis.com/go-build-log/5b7e8ca3/illumos-amd64_6c9b555f.log
> Finally some real test failure output! did this CL help that arrive?

Only to the extent that it uncorks the issue with the build uncovered with the new timer switch over, getting us back to the outstanding os/signal test failure.

I'm pretty sure the stack dump stuff is because with the tracing running in the background it's much more likely that TestAtomicStop will fail _a lot_, which kicks us up to the three minutes and the panic, rather than just a few failures which (though still fatal) fit within the timeout and are less explosive.

As an aside, I ran "all.bash" by hand with this change and (because the "os/signal" tests don't fail there) all tests appear to pass.

Patch-set: 1