crypto/x509: make loadSystemRoots load roots from multiple directories
"SSL_CERT_DIR" is meant to hold more than directory, when a colon
is used as a delimiter, but we assumed we'd only use a single directory
for all root certificates.
OpenSSL and BoringSSL properly respected that colon separator
as per:
* OpenSSL https://github.com/openssl/openssl/blob/12a765a5235f181c2f4992b615eb5f892c368e88/crypto/x509/by_dir.c#L153-L209
* BoringSSL https://github.com/google/boringssl/blob/3ba9586bc081f67903c89917f23e74a0662ba953/crypto/x509/by_dir.c#L194-L247

This change allows loadSystemRoots to load multiple roots
from various directories.

RELNOTE=yes

Fixes #35325

Change-Id: I0d554a00ccc34300a7f0529aa741ee7e2d5762f9