go/doc: add support for classifying examples
This CL is based on work started by Joe Tsai in CL 94855.
It's rebased on top of the latest master branch, and
addresses various code review comments and findings
from attempting to use the original CL in practice.

The testing package documents a naming convention for examples
so that documentation tools can associate them with:

	• a package (Example or Example_suffix)
	• a function F (ExampleF or ExampleF_suffix)
	• a type T (ExampleT or ExampleT_suffix)
	• a method T.M (ExampleT_M or ExampleT_M_suffix)

This naming convention is in widespread use and enforced
via existing go vet checks.

This change adds first-class support for classifying examples
to go/doc, the package responsible for computing package documentation
from Go AST.

There isn't a way to supply test files to New that works well.
External test files may have a package name with "_test" suffix,
so ast.NewPackage may end up using the wrong package name if given
test files. A workaround is to add test files to *ast.Package.Files
after it is returned from ast.NewPackage:

	apkg, _ := ast.NewPackage(fset, goFiles, ...)
	for name, f := range testGoFiles {
		apkg.Files[name] = f
	}
	p := doc.New(apkg, ...)

But that is not a good user experience. So a new function
NewWithExamples is added. It builds on top of the logic to
extract examples from Go source files offered by function Examples,
but also performs association of example to appropriate identifiers.

Examples fields are added to Package, Type, and Func. They are
documented to only be populated when NewWithExamples is used.

The new behavior is:

	1. NewWithExamples computes package documentation
	   while also accepting examples that were parsed
	   from test files.
	2. It assigns each Example to each Package, Type,
	   or Func.
	3. Malformed examples are skipped.

This change implements behavior that matches the current behavior
of existing godoc-like tools, and will enable them to rely on the
logic in go/doc instead of reimplementing it themselves.

Fixes #23864

Change-Id: Iae834f2ff92fbd1c93a9bb7c2bf47d619bee05cf