Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Thank you, Brad for the review!
> 
> I should have sent this CL earlier but I have been quite overwhelmingly swamped, so couldn't reply to the issue and mail out the CL early enough.
> 
> It came in on the day of the freeze though :) and Go1.14 could benefit from the conformance that this CL brings, plus it doesn't change a lot.

Even if this CL is super safe, that's not so much the point. It's just about respecting the release process.

Also, nobody had complained about this for like 10 years, so it can't be super important.

Feel free to appeal to Andy, though. I won't object if this does goes in (perhaps it's a bigger problem for something else?), but by default the assumption should be that it doesn't go in.

Patch-set: 4
Reviewer: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>