Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi, you marked this CL "DO NOT REVIEW".  How come?  Should we review this CL now?  Thanks.

I would be happy to send this out for review if you think it would be a reasonable addition.

This was supposed to be an example solution to https://golang.org/issue/33746. I was waiting to send this for review until it was decided how Go/QEMU wants to proceed. QEMU does not currently seem interested in accepting the SIGRTMIN+1 patch[1,2] mentioned here. QEMU's decision does not block this patch, but I'm not sure if Go would want to apply this if QEMU is not patched itself. It would only be useful to users with a locally patched version of QEMU. Though, in the long term, it may cause the maintainers to decide that the backwards-compatibility break is not as severe, since Go binaries were specifically called out[3] as one of the reasons the patch couldn't be applied.

[1] https://lists.nongnu.org/archive/html/qemu-devel/2019-08/msg05642.html
[2] https://lists.nongnu.org/archive/html/qemu-devel/2019-08/msg05950.html
[3] https://lists.nongnu.org/archive/html/qemu-devel/2019-07/msg00048.html

Patch-set: 1