cmd/compile: extract top level args at beginning of rewrite rule functions
This is a manual CSE of commonly-used values.
It reduces generated code and, in particular,
the number of bounds checks.
The most common case for rewrite rules is failure,
which means that the code executed before each rewrite failure
is the hottest.
This change reduces CPU time for compilation by approx 0.5%,
which (for this realm) is pretty good.

DO NOT SUBMIT

This change conflicts with CL 213703, which is probably higher priority.
It could be made to work with CL 213703 by using _i0 == 1 to swap
v_0/v_1 instead of load them.

Change-Id: I33f6dc6c126b208b97cf0f1e21039dace47e8aad