cmd/doc: show the package clause always
If no writes to the package buffer happen, then the package clause
does not get printed. This is a bug for cases where a file just contains
the package clause.

We fix this by making a dummy call to Write in the flush method which
gets called on all code paths. And remove the "len(p) > 0" restriction
to allow it to happen.

Updates #31457

Change-Id: Ia3bd0ea3963274c460a45d1e37fafc6ee0a197f0