Update patch set 3
Patch Set 3: Code-Review+2

This has been sitting here for a while; I like to get this in for 1.14 as the improvements are huge, at least for large numbers. I have verified the benchmarks on my own machine, and while there seems to be a slow-down around the threshold point, it seems it's probably mostly noise (I've moved the threshold to see variations). In any case, it would be trivial to move the threshold if need be.

I have not reviewed (or tried to understand) the algorithm in detail - but the overall code LGTM. I (or somebody else) should eventually take the time to review the algorithm. In any case, with algorithms such as this one, even the tiniest error would lead to crashes or test failures, so I am not too concerned.

Patch-set: 3
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2