Update patch set 4
Patch Set 4:

> Have there been benchmarks that we need an asm implementation? The assembler version looks like something a compiler can generate easily. What do I miss here?

That's a reasonable question.  We've gone down the path of using assembler implementations for all CPUs, but it's not clear that that is helpful.  It's clearly helpful on amd64, but for mips, riscv, probably others, it's not clear that we are doing any better than the compiler.

Patch-set: 4