Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Our usual preference is to avoid knobs.  What would happen if we simply did this unconditionally?
> 
> If this must be tunable, it should probably be done via a GOAMD64 environment variable along the lines of GO386, GOARM, etc.

The patch has two known side effects.

1. It generally slows down the assembler as additional passes are required if the assembler detects a jump that needs to padded.

2. It leads to an increase in binary size.  This is due to the extra bytes that are inserted in the code stream and the change in function alignment from 16 to 32 bytes.  There is typically more space between the functions in the final executable.  On the binaries I've been testing with I've seen increases of between 1.5% - 2%.

The GOAMD64 looks like a good option.  Ideally, I'd like to be able to be able to allow the user to specify the maximum number of prefixes that can be inserted before an instruction.

Would the following be acceptable?

GOAMD64 = "" (not defined) - The patch is disabled and the assembler works as it currently does
 
GOAMD64 = "pad_jumps" -  Jumps are padded so that they don't cross or end on 32 byte boundaries.  The maximum number of prefixes permitted per instruction is 5.

GOAMD64 = "pad_jumps:n" -  Jumps are padded so that they don't cross or end on 32 byte boundaries.  The maximum number of prefixes permitted per instruction is n, where n is a number >= 1 && <=5.

Patch-set: 1