Update patch set 2
Patch Set 2:

LGTM but we are in the freeze for Go 1.14 and only changes that improve stability or fix issues are permitted to go in.

While this is a trivial and correct change, it doesn't matter for stability, and there's no important a performance difference either (if any).

For Go1.15.

Patch-set: 2
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>