runtime:  static lock ranking for the runtime (enabled by GOEXPERIMENT)
I took some of the infrastructure from Austin's lock logging CR
https://go-review.googlesource.com/c/go/+/192704 (with deadlock detection from the
logs), and developed a setup to give static lock ranking for runtime locks.

Static lock ranking gives better runtime checking for possible deadlocks, since
you don't have to run into the deadlock -- you just have to run into a lock
acquisition sequence that violates the lock ranking. Also the static lock ranking
gives you a "memory" of lock ordering across many runs, rather than just a single
program run or a single run of all.bash. Along the way, I found a deadlock
involving the new timer code, which Ian fixed via
https://go-review.googlesource.com/c/go/+/207348

See the constants at the top of runtime/locklog.go to show the static lock ranking
that I ended up with, along with some comments. This is great documentation of the
current intended lock ordering when acquiring multiple locks in the runtime.

I also added an array arcs[] which shows and enforces the current partial ordering
among locks.  This is more specific about the dependencies among locks.

I don't try to check the ranking within a lock class with multiple locks that
can be acquired at the same time (i.e. check the ranking when multiple hchan locks
are acquired).

Currently, I am doing a lockInit() call to set the lock rank of most locks. Any
lock that is not otherwise initialized is assumed to be a leaf lock (a very high
rank lock), so that eliminates the need to do anything for a bunch of locks
(including all architecture-dependent locks). For two locks, root.lock and
notifyList.lock (only in the runtime/sema.go file), it is not as easy to do lock
initialization, so instead, I am passing the lock rank with the lock calls.

For Windows compilation, I needed to increase the StackGuard size from 896 to 928
because of the new lock-rank checking functions.

Checking of the static lock ranking is enabled by setting
GOEXPERIMENT=staticlockranking before doing a run.

To make sure that the static lock ranking code has no overhead in memory or CPU
when not enabled by GOEXPERIMENT, I changed 'go build/install' so that it defines
a build tag (with the same name) whenever any experiment has been baked into the
toolchain (by checking Expstring()). This allows me to avoid increasing the size
of the 'mutex' type when static lock ranking is not enabled.

Fixes #38029

Change-Id: I154217ff307c47051f8dae9c2a03b53081acd83a